+++ README.md
@@ -35,17 +35,17 @@ and workflow:
    The starting point is the client, represented here by a user making HTTP requests. This could be a browser, a mobile
    app, or any client capable of sending HTTP requests to the server.
 
-2. Router configuration ([`Router.java`](./src/main/java/com/fortun/backend/Router.java)):
+2. Router configuration ([`Router.java`](./src/main/java/com/fortun/backend/router/Router.java)):
    Upon receiving a request, the Spring Boot application uses RouterFunctions to direct the request to the appropriate
    handler based on the URL and HTTP method. This approach decouples routing logic from request handling, enhancing
    modularity and clarity.
 
-3. Request handling [`Handler.java`](./src/main/java/com/fortun/backend/Handler.java):
+3. Request handling [`Handler.java`](./src/main/java/com/fortun/backend/handler/Handler.java):
    The selected handler is responsible for processing the incoming request. It parses query parameters, validates
    inputs, and performs business logic, which includes calling on the repository interface to fetch or manipulate data
    stored in the database.
 
-4. Reactive repository [`PriceRepository`](./src/main/java/com/fortun/backend/PriceRepository.java):
+4. Reactive repository [`PriceRepository`](./src/main/java/com/fortun/backend/repository/PriceRepository.java):
    The repository interface defines reactive methods for database access, leveraging Spring Data's support for R2DBC.
    This allows for non-blocking I/O operations when querying the database, fitting seamlessly into the reactive
    programming model employed by the entire service architecture.
